feat: add faculty ui

+++ src/main/java/university/ui/FacultyMenu.java
@@ -4,7 +4,7 @@ import university.domain.Faculty;
 import university.domain.Person;
 import university.domain.Teacher;
 import university.repository.FacultyRepository;
-import university.repository.TeacherRepository;
+import university.repository.PersonRepository;
 import university.service.FacultyService;
 
 import static university.service.SearchService.*;
@@ -14,16 +14,16 @@ import java.util.Optional;
 public class FacultyMenu {
     protected final FacultyRepository facultyRepository = FacultyRepository.get();
     protected final FacultyService facultyService = new FacultyService(facultyRepository);
-    protected final TeacherRepository teacherRepository = TeacherRepository.get();
+    protected final PersonRepository teacherRepository = PersonRepository.get();
 
     protected void facultyManaging() {
-        System.out.println("\n*-Управління факультетами-*");
-        System.out.println("1 - додати факультет");
-        System.out.println("2 - змінити факультет");
-        System.out.println("3 - видалити факультет");
-        System.out.println("0 - повернутись назад");
         boolean status = true;
         while (status) {
+            System.out.println("\n*-Управління факультетами-*");
+            System.out.println("1 - Додати факультет");
+            System.out.println("2 - Змінити факультет");
+            System.out.println("3 - Видалити факультет");
+            System.out.println("0 - Повернутись назад");
             String inputLine = scanner.nextLine();
             try {
                 int input = Integer.parseInt(inputLine);
@@ -32,10 +32,10 @@ public class FacultyMenu {
                         addFaculty();
                         break;
                     case 2:
-                        deleteFaculty();
+                        changeFaculty();
                         break;
                     case 3:
-                        changeFaculty();
+                        deleteFaculty();
                         break;
                     case 0:
                         status = false;
@@ -57,31 +57,36 @@ public class FacultyMenu {
         String name = scanner.nextLine();
         System.out.println("Введіть коротку назву факультету");
         String shortName = scanner.nextLine();
-        Teacher dean = null;
-        boolean found = false;
-
-        while (!found) {
-            System.out.println("Введіть ідентифікатор декана");
-            String teacherId = scanner.nextLine();
-
-            Optional<Person> optionalPerson = teacherRepository.findById(teacherId);
-
-            if (optionalPerson.isPresent()) {
-                Person person = optionalPerson.get();
-                if (person instanceof Teacher t) { // Pattern matching
-                    dean = t;
-                    found = true;
-                } else {
-                    System.out.println("Ця особа не є викладачем");
-                }
-            } else {
-                System.out.println("Особу з таким ID не знайдено.");
-            }
-        }
-
+        Teacher dean = receiveDean();
         System.out.println("Введіть контакти факультету");
         String contacts = scanner.nextLine();
         return new Faculty(code, name, shortName, dean, contacts);
+    }
+
+    protected Teacher receiveDean(){
+        Teacher dean = null;
+        boolean found = false;
+        while (!found) {
+            System.out.println("Введіть ідентифікатор декана(залиште поле пустим, щоб додати пізніше)");
+            String teacherId = scanner.nextLine();
+
+            if(!teacherId.isEmpty()){
+                Optional<Person> optionalPerson = teacherRepository.findById(teacherId);
+
+                if (optionalPerson.isPresent()) {
+                    Person person = optionalPerson.get();
+                    if (person instanceof Teacher t) {
+                        dean = t;
+                        found = true;
+                    } else {
+                        System.out.println("Ця особа не є викладачем");
+                    }
+                } else {
+                    System.out.println("Особу з таким ID не знайдено.");
+                }
+            }else found = true;
+        }
+        return dean;
     }
 
     protected void addFaculty() {
@@ -90,36 +95,63 @@ public class FacultyMenu {
 
     protected void deleteFaculty() {
         boolean found = false;
+        boolean exit = false;
         Faculty faculty = null;
-        while (!found) {
-            System.out.println("Введіть ідентифікатор факультету, що треба видалити");
+        while (!found && !exit) {
+            System.out.println("Введіть ідентифікатор факультету, що треба видалити (введіть 0 щоб повернутись назад)");
             String facultyId = scanner.nextLine();
 
-            Optional<Faculty> optionalFaculty = facultyRepository.findById(facultyId);
+            if (facultyId.equals("0"))
+                exit = true;
+            else{
+                Optional<Faculty> optionalFaculty = facultyRepository.findById(facultyId);
 
-            if (optionalFaculty.isPresent()) {
-                faculty = optionalFaculty.get();
-                found = true;
-            } else
-                System.out.println("Факультет з таким ID не знайдено.");
+                if (optionalFaculty.isPresent()) {
+                    faculty = optionalFaculty.get();
+                    found = true;
+                } else
+                    System.out.println("Факультет з таким ID не знайдено.");
+            }
         }
-        facultyService.deleteFaculty(faculty);
+        if (!exit)
+            facultyService.deleteFaculty(faculty);
     }
 
     protected void changeFaculty() {
         boolean found = false;
+        boolean exit = false;
         String facultyId = null;
-        while (!found) {
-            System.out.println("Введіть ідентифікатор факультету, що треба замінити");
+        while (!found && !exit) {
+            System.out.println("Введіть ідентифікатор факультету, що треба замінити (введіть 0 щоб повернутись назад)");
             facultyId = scanner.nextLine();
 
-            Optional<Faculty> optionalFaculty = facultyRepository.findById(facultyId);
+            if (facultyId.equals("0"))
+                exit = true;
+            else{
+                Optional<Faculty> optionalFaculty = facultyRepository.findById(facultyId);
 
-            if (optionalFaculty.isPresent()) {
-                found = true;
-            } else
-                System.out.println("Факультет з таким ID не знайдено.");
+                if (optionalFaculty.isPresent()) {
+                    found = true;
+                } else
+                    System.out.println("Факультет з таким ID не знайдено.");
+            }
         }
-        facultyService.updateFaculty(facultyId, facultyGenerator());
+        if (!exit)
+            if (facultyRepository.findById(facultyId).get().getDean() == null){
+                System.out.println("Додати лише декана? y/n");
+                String choice = scanner.nextLine();
+                switch (choice) {
+                    case "y":
+                        Faculty faculty = facultyRepository.findById(facultyId).get();
+                        new Faculty(faculty.getID(), faculty.getName(), faculty.getShortName(), receiveDean(), faculty.getContacts());
+                        break;
+                    case "n":
+                        facultyService.updateFaculty(facultyId, facultyGenerator());
+                        break;
+                    default:
+                        System.out.println("Введіть коректне значення");
+                }
+            }
+
     }
 }
